created all functions

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: sshimizu <sshimizu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/11 03:43:49 by sshimizu          #+#    #+#             */
-/*   Updated: 2023/01/12 08:33:12 by sshimizu         ###   ########.fr       */
+/*   Updated: 2023/01/14 01:31:33 by sshimizu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
